Pinky: use uint read and write variants to work around rdar://39206631.

+++ Pinky/Pinky/PNKArgmax.metal
@@ -11,9 +11,9 @@ using namespace metal;
 constant half scale [[function_constant(0)]];
 
 template <typename T>
-void argmax(T inputImage, texture2d<half, access::write> outputImage, ushort2 gridIndex,
+void argmax(T inputImage, texture2d<half, access::write> outputImage, uint2 gridIndex,
             ushort featureChannels) {
-  const ushort2 outputSize = ushort2(outputImage.get_width(), outputImage.get_height());
+  const uint2 outputSize = uint2(outputImage.get_width(), outputImage.get_height());
   if (any(gridIndex >= outputSize)) {
     return;
   }
@@ -35,13 +35,13 @@ void argmax(T inputImage, texture2d<half, access::write> outputImage, ushort2 gr
 kernel void argmaxSingle(constant ushort *featureChannels [[buffer(0)]],
                          texture2d<half, access::read> inputImage [[texture(0)]],
                          texture2d<half, access::write> outputImage [[texture(1)]],
-                         ushort2 gridIndex [[thread_position_in_grid]]) {
+                         uint2 gridIndex [[thread_position_in_grid]]) {
   argmax(inputImage, outputImage, gridIndex, featureChannels[0]);
 }
 
 kernel void argmaxArray(constant ushort *featureChannels [[buffer(0)]],
                         texture2d_array<half, access::read> inputImage [[texture(0)]],
                         texture2d<half, access::write> outputImage [[texture(1)]],
-                        ushort2 gridIndex [[thread_position_in_grid]]) {
+                        uint2 gridIndex [[thread_position_in_grid]]) {
   argmax(inputImage, outputImage, gridIndex, featureChannels[0]);
 }
