+++ Pinky/Pinky/PNKGammaCorrection.metal
@@ -10,7 +10,7 @@ constant float gamma [[function_constant(0)]];
 
 kernel void gammaCorrect(texture2d<half, access::read> inputImage [[texture(0)]],
                          texture2d<half, access::write> outputImage [[texture(1)]],
-                         ushort2 gridIndex [[thread_position_in_grid]]) {
+                         uint2 gridIndex [[thread_position_in_grid]]) {
   if (gridIndex.x >= inputImage.get_width() || gridIndex.y >= inputImage.get_height()) {
     return;
   }
